reorganise snake lib into sub-directories
